requirements.txt
python-telegram-bot==20.6
cloudscraper==1.2.71
pytz
requests
python-dotenv
aiohttp==3.9.3
brotli